Check for the note owner where an owner signature is needed
Client::account refused any signer that was not the note owner, so no
session could hold a divergent pair. Spending never needed the owner:
pool.rs builds transact from the signer address, and the contract asks
only for sender.require_auth() plus the funding transfer. The check
comes off session open, and a delegated session now opens.

Key derivation and registration do need the owner, and now check for
themselves. The derivation signature is the note secret and WalletSigner
holds one account, so a divergent pair would derive the payer's keypair
and file it under the owner's address; the web client checks before the
wallet is asked for anything. Registration's auth entry is keyed to the
owner, which the payer cannot fill. An owner that has never derived keys
therefore still opens the first session itself.

SignerIsNotNoteOwner named the session rather than the step, so reword
it. The web guard tests were plain #[test] under the wasm32 harness and
never ran; they use wasm_bindgen_test now.
