+++ sdk/native/src/error.rs
@@ -29,15 +29,17 @@ pub enum Error {
     #[error("wallet request rejected by user: {0}")]
     UserRejected(String),
 
-    /// The signing account is not the note owner.
+    /// An operation needs the note owner's own signature, and the session
+    /// signs with a different account.
     ///
-    /// The two transaction paths disagree on which identity drives them:
-    /// transact sources the envelope, `sender` and sequence number from the
-    /// signer, registration from the owner. Which is right when they differ is
-    /// unsettled, so a divergent pair is refused rather than resolved here.
+    /// A session may hold a divergent pair: the payer sources the envelope,
+    /// `sender` and sequence number, and spending needs nothing else. Two
+    /// operations do need the owner itself — key derivation, whose signature
+    /// *is* the note secret, and registration, whose simulated auth entry is
+    /// the owner's. Those raise this; the session as a whole does not.
     // Escapes to a UI toast, the telemetry ring buffer and CLI logs.
     #[error(
-        "signing account {} is not the note owner {}; a session where they differ is not supported",
+        "signing account {} cannot sign for the note owner {}; this step needs the owner's own signature",
         crate::types::Sensitive(signer),
         crate::types::Sensitive(owner)
     )]
